Fix bug in Makefile, added matlab script
The libso target of Makefile worked fine on a mac but I needed -lmarbl at the
end of the gfortran call (rather than coming before compiler flags). Just to
have a note of it somewhere:

$ nm marbl_lib/marbl_interface.so | grep vtab_marbl_interface

on the mac, the Marbl_interface_class symbol was classified with S (section
symbol) but on cheyenne it was U (undefined symbol). Now cheyenne shows it as D
(initialized data) and the python script runs.

Also, I combined the mex calls for initialization into a matlab script; note
that this currently causes a seg-fault, hopefully just due to a threading issue
that will be easy to clear up.