More Makefile clean-up
Separate the compiling of marbl_interface_wrapper_mod.F90 from the construction
of the shared library.

Also, the portion of the last commit message concerning the matlab script was
incorrect -- it will not cause a seg-fault because I wrapped the calls in a
spmd(1) block to force it not to be threaded.