init_marbl() Fortran routine needs domain info
Instead of constructing a generic column in the marbl_interface_wrapper_mod.F90
module, the delta_z, zw, and zt arrays (and their size) are intent(in).

I updated the python interface to generate the same generic column that used to
be hard-coded in the Fortran, but have not yet updated the mex code so the mex
driver will probably not build (and if it does build, it definitely won't run).

I have also cleaned up the python code quite a bit:
1. Adding lots of comments to help users decipher what is going on
2. making the init_marbl() routine return both a status code AND nt
   instead of trying to return nt as an argument