Mex interface for init returns log to driver
We were having issues with mexPrintf and strings that contained special
characters (for example, the '%' showing up in all our derived types). In
trying to work around that, we decided that it made the most sense to have the
mex driver return an array of strings representing the log to matlab, and
matlab can push it through fprintf.

I still need to add this log output to non-init calls in the mex driver, as
well as the python driver. I also need to return something for marbl_log when
init fails (replacing another mexPrintf call).